rename: 반복주기 상세 DTO 패키지 이동
