+++ src/main/java/project/volunteer/domain/recruitment/application/RepeatPeriodService.java
@@ -2,7 +2,7 @@ package project.volunteer.domain.recruitment.application;
 
 import project.volunteer.domain.recruitment.application.dto.RepeatPeriodParam;
 import project.volunteer.domain.recruitment.domain.Recruitment;
-import project.volunteer.domain.recruitment.dto.RepeatPeriodDetails;
+import project.volunteer.domain.recruitment.application.dto.RepeatPeriodDetails;
 
 public interface RepeatPeriodService {
