+++ src/main/java/project/volunteer/domain/recruitment/mapper/RecruitmentFacade.java
@@ -17,7 +17,7 @@ import project.volunteer.domain.recruitment.application.dto.RecruitmentParam;
 import project.volunteer.domain.recruitment.application.dto.RepeatPeriodParam;
 import project.volunteer.domain.recruitment.domain.Recruitment;
 import project.volunteer.domain.recruitment.domain.VolunteeringType;
-import project.volunteer.domain.recruitment.dto.RepeatPeriodDetails;
+import project.volunteer.domain.recruitment.application.dto.RepeatPeriodDetails;
 import project.volunteer.domain.sehedule.application.ScheduleService;
 import project.volunteer.domain.sehedule.application.dto.ScheduleParamReg;
 import project.volunteer.domain.user.application.UserService;
